+++ Src/CheckMade.Common.Model/Core/Geo.cs
@@ -5,4 +5,4 @@ namespace CheckMade.Common.Model.Core;
 public record Geo(
     Latitude Latitude,
     Longitude Longitude,
-    Option<float> UncertaintyRadiusInMeters);
+    Option<double> UncertaintyRadiusInMeters);
