+++ Src/CheckMade.Tests/Unit/ChatBot/Logic/InputProcessorTests.cs
@@ -206,7 +206,7 @@ public class InputProcessorTests
         var inputGenerator = _services.GetRequiredService<ITlgInputGenerator>();
         var locationUpdate =
             inputGenerator.GetValidTlgInputLocationMessage(
-                new Geo(17, -22, Option<float>.None()));
+                new Geo(17, -22, Option<double>.None()));
         
         var serviceCollection = new UnitTestStartup().Services;
         var (services, container) = serviceCollection.ConfigureTestRepositories();
